ClearCommand CameraShakeCommand

+++ src/main/java/cn/nukkit/command/tree/node/FloatNode.java
@@ -10,14 +10,13 @@ import cn.nukkit.api.Since;
  */
 @PowerNukkitXOnly
 @Since("1.19.50-r4")
-public class FloatNode extends ParamNode<Double> {
+public class FloatNode extends ParamNode<Float> {
     @Override
     public void fill(String arg) {
         try {
-            this.value = Double.parseDouble(arg);
+            this.value = Float.parseFloat(arg);
         } catch (Exception e) {
             this.error();
         }
     }
-
 }
